rest: do not allow to specify the entity name
It's automatically assigned by the API. That allows to be sure that the
name is an UUID, so we can eventually optimize its storage/indexing.
Also, we make sure that the entity name stays a secret by being random,
and that no one can access it without knowing it explicitly.
This also simplify the case where there would have be a conflict, since
that's not impossible.