indexer, rest: simplify metric model
This change makes the metric to not be a 'resource' anymore. The
motivation behind that is simplification of the metric attribute to the
bare minimal, and the reduction of the global resource table (making
querying faster).

It will also lead to RABC simplification on the metric/resource
association later, as we'll be able to grant access to a metric if its
associated to a resource with the correct user_id/project_id (that is
not yet implemented).

Change-Id: I24ddd20f72746e8d64a4ae30f8132afb3efecaac