Partially revert database tests patches.
This partially reverts commit 01fb4f0af450a07619c26a72011c9a564b8b0bd7
and 5b9b43b88015348ef6816eba73283f7e5a1cc2fb

The first commit slowed down the test by a factor of 4, which turns out
be too much to my taste.
The second one is buggy as it tries to use a directory for locking that
is unique to a process, but since we run tests in parrallel, we need a
unique directory for ALL the process, so it has to be passed down as a
env variable.

Change-Id: I3fcccef0915eace1b23297d38d1a0a4212a13ff3