rest: fix exception JSON encoding
When an error occures, most of the code rely on pecan.abort(code, exc)
to return an exception to the HTTP client.

The problem is that the `exc' passed is given as is to the JSON encoder
that Pecan uses, and this one does not know how to encode an exception.
So we have to make sure a string representation of that exception is
sent down the wire, not the exception itself.

Change-Id: I6f9b9f0dd84c31726a82a5c17bee87c247eefd0e