Update resource.yaml to reflect recent fixes
Some bugs were fixed in previous commits with how data is presented at /
and /v1/resource but the xfail-ing tests in resource.yaml didn't get
fixed up because they were still failing (instead of becoming unexpected
success) because of the structure of the tests. We'll want to keep this
in mind when creating future xfails.

Change-Id: I7e301d3a97099480edd4ed13709c5865cd8b4929