Return detailed metric info on create
When a POST is made on /v1/metric return the full
metric object info instead of just id and archive
policy name like we do now. The is so the metric
api is consistant with the rest and the behavior
is now different as in we allow an optional name
so more info returned would be helpful.

Change-Id: I66681550b0534e6dfcd7fecee80ba01bf6cc73da