ceilometer: Use http session in dispatcher
The ceilometer dipatcher needs to do between 2 to 5 http requests
for each samples. Current it use the basic API of the python-requests
library, that creates 1 tcp connections for each http requests.
Connections are pooled by default (10 simultaneous connections) and
a error is raised when the pool is exhausted.

This change move to the session API of python-requests to
creates only one tcp connection per ceilometer-collector workers,
and block when the pool exhausts.

Change-Id: I8255622cf57082785b48754d9e506413ee9122a7