Correct use of phases in devstack plugin
Configuration of Ceilometer and Heat needs to happen in the
"post-config" phase not the "extra" phase, otherwise services (such
as the ceilometer-collector) will start without the correct
configuration, leading to unexpected behaviors and errors.

Ceph is moved as well for parallelism, even though it doesn't
matter: Ceph starts very early in the devstack process because it
needs to be running to be configured.

Change-Id: I3d7353049cf4898ad69159b73e590c78302c0c18