Cover and correct unicode handling in mod_wsgi setup
On some setups the default encoding that will be present in the
Python running under mod_wsgi will not be UTF-8 compatible and will
cause errors when decoding headers in the WSGI environment.

This changes the apache config template to force the mod wsgi to use
UTF-8 thus allowing werkzeug to properly decode URIs which are
urlquoted version of utf-8 encoded strings.

Change-Id: I6f470f8ab311f8debd1758840dd74b056abbc9ea