Ensure the indexer is disconnected in gabbi fixtures
An index.connect() was added in commit 0620d841 but no corresponding
disconnect(). There's evidence that not properly disposing of the
associated engine facade may be causing test failures when
sufficient gabbi fixtures end up in the same process (this requires
a relatively low number of available cores).

Even if the lack of disconnect is not the cause of that, it is still
a correct thing to have.

Change-Id: I53d6ffbdead5ee959521131aecb2d7b58b6f8dd6