pass LANG into testing environment
In the tox >= 2, the environment is restricted. In Python3 tests if
we do not pass LANG with a passenv command then encoding routines
which rely systemencoding settings will fail.

We could explicitly set a LANG but given that until now the tests
were relying on the environment's LANG this seems okay.

Change-Id: Ifd575538012ab3b47c5aaf00860a2f292dcb9c3e