Handle indexer connections more cleanly in tests
In some situations (especially low concurrency) it was possible to
run out of available postgres connections during test runs. This was
happening because the gabbi tests call gnocchi.rest.app:setup_app
with every test and were doing it without passing an indexer
reference. This caused setup_app to create a new one.

Now, instead, the gabbi tests flex some globals to insure that an
indexer, per test suite, is passed to setup_app.

setup_app is changed to _only_ call index.connect() when a new
indexer is created, making the assumption that if callers have
provided an indexer, they have also connected. This assumption is
born out by the tests.

The alembic online migration called disconnect() for cleanliness
(because it creates and connects its own indexer).

With these changes the tests will pass with only 6 available
postgresql connection running with a concurrency of 2, so the -N 200
has been removed.

Change-Id: I00430a45ab471b70bc747b9811ec984ab1493eaa