Introduce gnocchi-metricd
This daemon is responsible for treating new measures that are sent to
storage drivers and must be run in background.

Also add a thread in the gabbi ConfigFixture to do metricd
processing. Without this the new async processing of measures will
cause tests that check to see if measures were created will fail.
Meanwhile gabbi >=0.100.0 provides a 'poll' function so that tests
will try a few times for what they are looking for.

The tests in resource.yaml have been cleaned up to insure that they
are testing a specific resource and its measures, not an ambiguous
last-in-the-list member of resources. async.yaml confirms the async
handling.

The 'file' storage driver has been adjusted to make sure that writes
of aggregation files are atomic (by writing to a tmp file that is
renamed. Without this it is possible for files to exist but be read
as being empty.

Co-Authored-By: Chris Dent <chdent@redhat.com>
Change-Id: I0fd449ee82073edafca7673d00c69e61f75bb7d0