Don't reclone the repo we already did
In order to use a devstack plugin at all, we first must have cloned
the gnocchi repo pointed to in local.conf. This sets the remote and
the chosen branch. install_config then goes on to clone again,
clobbering the thing that just got checked out, including and
customizations that are being tested.

So, just get rid of the redundancy.

Change-Id: Ia60273ca73a351ba01f335ff79d5964ae3ddd6be