Ensure location header urls account for script_name
The API is at /v1... but if is mounted at /gnocchi/v1 then
"/gnocchi" is the script_name and we need to account for that when
generatin URLs for location headers.

Pecan already does this where the application_url attribute is being
used (for generating informational urls in response bodies).

A new gabbi test suite is added which runs all of the existing gabbi
tests with a prefix. Getting this to work identified a bug in gabbi,
thus the update to test-requirements.

Change-Id: I6e3b461610ad1b694e6ae3df33831b3f6224668f