Cleanup handling of metricd and its logs in devstack
We want to start it in the gate (otherwise async processing
is difficult) and we want to save logs to the places where
devstack expects logs to be. The primary way to do this is
for stderr to be captured in the run_process handling.

This removes the special GNOCCHI_LOG_DIR (which makes sense for a
regular installation but not a devstack one) so the run_process for
gnocchi-api needs to be changed as well.

Change-Id: I2f07143724cc62644aee65ecdde0b01212b2cfc7