Don't allow duplicate timestamps in carbonara series
If the calling code provides timestamp/value pairs with duplicate
timestamps pandas will raise a ValueError when trying to operate on the
index of the data set. We avoid this problem by removing duplicates in
the timeseries. The assumption being made here is that some single
metric can't have two different measurements at the same time, that's a
violation of physics.

The change is in place for both __init__ and set_values to insure
that the internal representation of the timeseries is consistent no
matter how it is created.

Change-Id: Ib293714e26eca6b5de7cd4a2307569ab3d8a9c43
Closes-Bug: #1475329