statsd: stop mocking indexer.get_metrics in tests
This is actually causing an issue because by mocking, we return a metric
that has not its archive policies eagerly-loaded, so that fails badly.
We need list_metrics() to have the .archive_policy eagerly-loaded as we
expect. There's actually no harm to not-mock that call, so let's use the
real call! This fixes the transient issue we saw with
DetachedInstanceError poping up sometimes when the attached session is
collected.

Change-Id: I3991cfb8b828be0ea51923ec5e6dcdc84a4d9b79