archive_policy: validate agg methods values
This creates a custom option that validate the input for the valid
default aggregation methods used in archive policies.

PS: Unfortunately, the oslo.config generator is dumb and does not know
the type of option so it has trouble generating that correctly, but it
still works anyway and it's included in the sample config.

Change-Id: Ie1f9378063d37607e81f400f6a7d506f6a4954ed