storage/carbonara: add timestamp as measure suffix
This patch adds a timestamp as a suffix to the name of recorded measures
for a metric. This does not change anything on how the code currently
works, it's just a way for the curious developer or operator to have an
easy glance on e.g. how behind gnocchi-metricd is on processing new
measures.

Change-Id: Id4916e6aabfe09433eabc7fa405ecdd5da9a3b53