Introduce a resource id translation shim
In the openstack environment resource ids are a mismash of UUIDs
(yay!) and other things (boo!). When gnocchi receives a resource id
that is not a UUID, it drops it on the floor with a 400 error.

In some sense this is the exact right thing to do. In others it is
entirely impractical given the use context.

This change introduces a shim which accepts any string as a resource
id, either in the URL or in the Resource voluptuous Schema and does
a transformation on it to turn it into a UUID. That transformation
is:

* check if the string <= 255 in length (ValueError if not)
* create a uuid5 using RESOURCE_ID_NAMESPACE UUID and the string
* use the resulting UUID as the local id for the resource

This is a recommended form of creating URNs via rfc 4122
https://www.ietf.org/rfc/rfc4122.txt

The transform is done for both writes and reads. This is necessary
because we have no control over who or what is the the caller using
the API. Gnocchi users will expect to be able to access data in the
Gnocchi API by the IDs they are familiar with.

Doing things this way:

* keeps UUIDs as the primary identifiers internally
* requires no indexer-database migration (which a change in id
  structure would).
* is less intrusive than providing a persistent map between
  OpenStack key and Gnocchi key.
* may raise awareness about the value of using UUIDs everywhere

Change-Id: Ibfc58d6f239a0474cbb3ccd9a6229393ada77e8f
Related-Bug: #1480346