In test_rest run process_measures after a request
When done before we are not guaranteeing that the request just made
is processed in a way that a subsequent assertion expects.

Presumably the previous way was done to say "make sure the previous
request was fully processed before doing this request, which is a
confirmation". However we have seen intermittent test failures where
expected measures are not present in result sets.

Change-Id: I67a0e24195b3fe2d8406495fedd072307b552990