Removing ceph locking system
The ceph locking system doesn't work as expected.
When the client loose the connection to ceph, Locks are
not released, so more code need to be added to expire
the lock when it's no more updated and/or break it when
we detect a client that hold the lock is dead.

Instead of this, this change just use tooz as meachnism system.

And a ceph driver will be added to tooz instead. Tooz have
already a tests coverage for drivers, it should be easier to
detect this kind of race.

This change have a migration impact, changing the lock meachism
means all gnocchi daemons must be stopped before upgrading them.

Change-Id: I5662817db9cd6bbb7dc220407df42c55f862ef6b
Closes-bug: #1500646