Make sure that swift doesn't block gnocchi startup
For reasons that are still not entirely clear, when running gnocchi
with swift as the storage backend, there is an apparent race
condition that is alleviated by making two changes:

* starting gnocchi-metricd after we have confirmed the api server is
  up
* looping our test to see if the api server is up with a --max-time
  setting to handling blocked requests

In the previous code if a session was blocked it would consume the
entire while loop and no retry would happen.

We'd like to know more about why this is happening but for we also
want to get the gate working. This is a start.

Change-Id: Iafa4e6dc7d4a6bada4c3e81424a470be3d015d2a