carbonara: fix exception when creating metric
_create_metric() never returns MetricDoesNotExist, which makes no sense.
But it can raise MetricAlreadyExists.

Now, the way we use this function is always inside a lock that is tied
to the metric, so it's impossible that this exception is raised in
theory – and we never saw it so far in practice.

But still, for the sake of it, let's fix that mistake!

Change-Id: I256c3d6e7a0c4ac1c50af145b1938106289bce88
(cherry picked from commit b810d2252418b7ad617e4d9d9aeb6de52ddfbef6)