fix expunge_metrics method
expunge_metrics method does not work properly. it uses list_metrics
from indexer to get metrics with status == 'delete'. this makes sense
but list_metrics itself is hardcoded to always search for 'active'
status. because of that, the query will always be look for status ==
'active' and status == 'delete' which can't happen so it never finds
any match.

Change-Id: I318fa5b4642a187776dce23ee9a933a5f852c4f8