sqlalchemy: fix metric expunge
There's a typo in the current driver which prevents expunging the metric
for real. This patches fixes that and ensures the code works.

It also deletes a check in another test that might fail due to the fact
that expunge can be run at the same time by another test process.

Change-Id: I2d02b338bb5d98411a097ecba22d6934bd61ed3b