ceph: fix the metric list to process with new measures
Currently, the list returned in the Ceph driver contains a lot of
doubloons because it returns a list and not a set. If 1 metric has N new
measures batch waiting to be processed, the list return will be size N
and not 1.

Using a set() avoids that issue and the memory draining implied.

Closes-Bug: #1533793
Change-Id: I3a0b726aae14a17a23a365babc1a2537fb4d1052