Skip already processed measurements
When using multiple workers, the list of metrics to process
could possibly include those that have already been
procesed by other workers.  This causes the worker to do
extra processing.

This patch addresses this by ensuring that there is
measurements in the system first before starting to
process it.

Change-Id: I0307d3ca679c087d141db6bd02d8de1706298bbb
Partial-Bug: #1536909