Disable rados thread system
This change disable rados thread system, that just slow down metricd.

A quick example of performance gain: http://paste.openstack.org/show/485202/

The side effect of removing this thread is that a SIGINT during a
Ceph API call will be handled only when the call is ended.

https://github.com/ceph/ceph/commit/2ce5cb7d06c80e34824981c36adcaa2dbfbc581f

Future version of python-rados should move to cython instead of ctypes
and should not thread and suffer of this issue.

Change-Id: I3dfc04f49df0129eaaa3f1ebb1917efbea9b5355