tests: remove measures reporting test
While it's very nice to test it, the test is unreliable: it works by
chance 99% of the time. Indeed, we have other tests running on the same
storage driver with another metricd-like process running, and these
process can process/delete our unprocessed measures at any time, making
this test fail as the output from measures_report() will differ.

As the function is used in gnocchi-metricd, let's assume it's tested as
part of the functional tests.

Change-Id: Ie1ceddec075e1cb86759596957da153540cf433c