carbonara: remove carbonara-dump
This tool was the last man standing of our tool, and it's now time to
remove it. We dropped the -create and -update command since 1.3, and
since the format changed and we deprecated TimeSerieArchive, it's now
time to clean that old stuff completely.
Considering the format change, we would need an entire new set of tool
at this stage anyway.

Change-Id: Iecba85ab70695eb9eb8bb31386b1bec86457f0f6