minimise calls when generating report
in swift and ceph backends, we make a separate call for every single
metric to gather the measures for each metric specifically. we
don't really need this and some of the output doesn't even make it
visible. this also might be thousands of calls depending on backlog
size.

this patch adds a high-level summary equivalent to what we currently
show and a details param to display full per metric details.

Change-Id: I3e8d887e22d4941673a658f60c3b3c4816c51133
Closes-Bug: #1544081