Make the content of PKG-INFO can be decoded
The README.rst contain a dash which is \xe2\x80\x93, the PKG-INFO file
will generated by the content of README.rst when install gnocchi package.
And the pbr will read the PKG-INFO file to get package info, in py34
environment, the dash cannot be decoded and will raise an error.

Change-Id: I6a3c85beb260115e7ad32bfe101ced25442aa8d1
Closes-Bug: #1553893