Increase partitioning logic period
gnocchi uses the reporting worker to partition tasks across each of
the processing workers. by default reporter runs every 10s and
processor worker every 5s. unfortunately, in any environment, small
or large, the processor takes much longer than 10s to process it's
data. because of that, the repartition messages that the reporter
sends up queues up so the processor ends up running through a list
of stale repartition messages before it can grab a new set of
metrics.

This calls for an increase in the period of which the reporter runs.

Change-Id: Ie7b114a1c8f55e7291057e8193172c4f188a2f61
Closes-Bug: #1554738