disable swift retry
swift has native retry functionality built in and it's enabled by
default. this means every action we do, even those expected to failed,
will repeat 5 times with an additional backoff mechanism. so every
time we have a new metric, there is quite a large delay between first
attempt and confirmation that it does not exist

Change-Id: I4320fb3e8bb30603bd70f8159fbcf855fc4a2880