Tuneup gabbi archive_rule.yaml file to modern standards
Make use of METHOD: /url shorthand to make the request method more
visible and more tightly associated with the url for easier reading.

In one place a status was move above a response_headers so that the
test more clearly aligns with the order of an HTTP request. This
isn't required by gabbi, but makes reading nice.

Change-Id: I37afcb780db8e9491fab5b03165ab1d8adccd0d2