Tuneup gabbi archive.yaml file to modern standards
Make use of METHOD: /url shorthand to make the request method more
visible and more tightly associated with the url for easier reading.

Use $LAST_URL to refer to the URL of the previous request, but only
when the sameness of the URL is a relevant part of the expression of
the tests-in-sequence. (That is, don't just use LAST_URL because the
URL is the same.)

In one test use more complex json paths to confirm the details of
the list of archive policies.

Change-Id: Ic0b02b19bea121af9c21157443d0f1233b5f4e0c