Tuneup gabbi history.yaml file to modern standards
Make use of METHOD: /url shorthand to make the request method more
visible and more tightly associated with the url for easier reading.

Use $LAST_URL to refer to the URL of the previous request, but only
when the sameness of the URL is a relevant part of the expression of
the tests-in-sequence. (That is, don't just use LAST_URL because the
URL is the same.)

Change-Id: I17d980830c114db14cc0881798a35f0c46bce3fa