Add http_proxy_to_wsgi to api-paste
This sets up the HTTPProxyToWSGI middleware in front of Gnocchi. The
purpose of thise middleware is to set up the request URL correctly in
case there is a proxy (For instance, a loadbalancer such as HAProxy)
in front of Gnocchi.

So, for instance, when TLS connections are being terminated in the
proxy, and one tries to get the versions from the / resource of
Gnocchi, one will notice that the protocol is incorrect; It will show
'http' instead of 'https'. So this middleware handles such cases.
Thus helping Keystone discovery work correctly.

The HTTPProxyToWSGI is off by default and needs to be enabled via a
configuration value.

Change-Id: Ic5526cf37e70335fa2cc70946a271253f227f129
Closes-Bug: #1590608
(cherry picked from commit 2b1ed8d4528df8c3071343d1cd5764b6a7122fd1)