rest: don't use is_body_seekable
The condition was wrongly reversed in previous attempt to
fix the wsgiref hanging bug. Making the change useless the code
was always returning an unseekable version of the wsgi.input.

Also using is_body_seekable is useless body_file_seekable already do
this check (correctly) and return the raw one when the wsgi input is already
seekable.

Change-Id: Ifdfd419d964c8cfc9295bbf1c772429c68e5618d