archive_policy: lighten the default archive policies
This reduces by one or two thirds the default archive policies CPU
consumption for 'low' and 'medium' by removing one or two definitions.
Also make it clearer that using low and medium is going to be fasten in
terms of CPU as the consumption depends on the number of definition in
an archive policy.

Change-Id: Iaba3b2ef88858ad777147d2859180d9a27658f1c