allow required=True when patching resource type
We have currently two limitations about resource type update.

* We cannot set required=True on a new attribute
* We cannot update a attribute

This change introduces the notion of 'options' for the description of
a new attribute or an attribute to update.

And implements the first item on top of this "option".

The option added is 'fill' to known to fill all row of existing
resources with a default value.
And 'fill' is required if 'required' is set to allow new
attribute with required=True.

Change-Id: If0bd609ed586b6fbe4fe7877ece237e55baa7d45