carbonara: Don't use clean_ts()
Checking if the index have duplicate have a high cost. And everything is
already ordered and unique. So we don't need to do this on each
timeserie creation.

We have two place where we need to order the series. When new data are
added (set_values()), when we build the object from v2 format (from_data()).

This increases the perf ~33% for each kind of timeseries objects.

Change-Id: If530df7a3da07555eb6d8447fa17c72adc6015a6