tests: split two kind of functional tests
We have two kinds of functional tests. Those run with
the wsgi_intercepter and those run on a real Gnocchi server.

We currently the wsgi_intercepter tests in two jobs, the classic tox
pyXX and the functional one.

This change rearrange tests to makje clear the purpose of each tests
and to not run them twice in gate.

Change-Id: Ie0339ed92b790bebb0f4a5ed0dc5b690fa8ce4fa