carbonara: remove numpy warning
We ask scipy/numpy to compute std on serie with one timestamp. And numpy
complains and return nan. Next we drop all nan.

This change creates the list of timestamps that are not unique and then
compute the std. numpy does not complain anymore and we don't have to
remove nan.

Closes-bug: #1663419
Change-Id: Ie1ddd244e42dcac4f91f741f37a980e452f91845