Correct bad use response_strings in live.yaml
response_strings wants a list, not a string. In gabbi earlier than
1.33.0 it accept a string and then iterated checking for each single
character in the response body. So still passing.

This fixes recent failures by putting the necessary '-' at the start
of the search string. It's an easy bug to make and then never
notice, which is why gabbi 1.33.0 is all uptight about it.

Change-Id: I02042c0b5005a50b3760c9aae67ad4d0eeb69e04