Use NULL as creator in ResourceUUID conversion
In older versions of Gnocchi, the creator could be null. It's unlikely now as
all auth helper returns a string as creator, but it's still possible in theory.
This patch uses the NULL string as the creator is it's None (so NULL in SQL).

Change-Id: Id1faa16512b995fa2cd24ca4bed7912934cb5aad
Closes-Bug: #1684246
(cherry picked from commit 75e420673159fcf4a383f130e1652243fccc2b75)