ceph: store measures in omap values
ceph sucks for small objects; POSTs to ceph incoming are always
small objects.

this changes logic to store new measures as omap values instead of
objects. this allows us to write to 'memory' aka leveldb/rocksdb
instead of disk.

this does not change the durability agreement since we are already
storing object keys in omap so if omap fails, we will lose link
to objects regardless if on disk or not.

using local 20OSD ceph cluster, with 18 metricd. this is:
- ~2x faster than aio_write patch to POST
- ~2x faster than aio_write patch to process
- ~3x faster than no aio_write patch to POST
- ~3x faster than no aio_write patch to process
- ~3.5x faster than 3.1 to POST
- something a lot faster than 3.1 to process (no idea why)

Change-Id: I4bae365955fdbafe4ad837596490774c42bc5251