reuse storage coordinator
the processing worker creates it's own connection to coordinator
to handle hashring distribution. this is unnecessary as the storage
driver already has a coordinator for sack_locking which is exactly
what the processing worker's coordinator is ultimately for. reuse
the same coordinator since if the storage coordinator is down, the
processing worker is ultimately down.