storage: fix resample on empty metric
If a metric is empty, carbonara raises an IndexError when checking for the data
on grouping since the Pandas timeseries is empty.

This is mostly hit in version >= 3.1 since 3.0 does not support resampling, but
you never know.

Fixes #69

(cherry picked from commit 5fb48769440cc405d375d7721fecf82e0dd5610f)