tests: Update travis/tox configuration
Travis ressources for opensource project is limited

This change will limit the number of jobs per PR in parallel, allowing
more PR to be processed at the same time.

And remove the build of the image of each commit (save 3-4 minutes).
It pulls the image from dockerhub (I will create a new repo to have
this image updated automatically).

It's run in parallel tests of differents storage system.

The overall time of testing is still ~40 minutes, but now we use only 3
jobs instead of 10.