Add a notification mechanism for getting sacks to actually process
This adds the ability for incoming storage to give hints on which sack should
be processed, rather than iterating on all sacks on every run.

If no hints are given for a `metric_delay_processing`, the sacks are all
iterated and checked anyway. Otherwise, only sacks with new measures are
processed.