ceph: change default timeout
The default ceph timeout is unlimited. For metricd this
is not really an issue. But for API, that makes connection
stuck until ceph comeback.

This change uses a default to ensure we return an error 500 by default,
and not hang.

This also change the type on the option, rados except string.

(cherry picked from commit 5f634d94ed593466f4fcca2f22376a468ff76986)