metricd: use a better number of partitions for the hashring
The current hashring distributing the sacks across the metricd workers is 200.
That means the the smallest number of sacks that a worker can have is
(number_of_sacks/(200*number of workers))=0.64 on a default installation with 1
worker.

That's a weird ratio.

If the number of sacks created is 4096 and the number of workers is 4096, then
each worker manages at least 4096/200=20,48 of sacks. That's weird. Each one
should manage at least one sack + replicas in such a case.

Let's use the number of sacks as the partition number so the minimum amount of
sack that a worker will always be able to manage is 1 (+ replicas).