ceph: Remove the async read of incoming measures
We don't have clue for now, but when rados_osd_op_timeout is set,
aio_read seems to return corrupted data or the add_measures callback
does not work as expected.

This change moves back to the synchronous read of the measures like in
3.0.

This code is gone in 4.0 and later, because we use omap attributes
instead.