Only retry connection to external components in metricd
Currently storage such as Swift will retry forever to connect on __init__.
Actually, the only process that wants to retry indefinitely is metricd. The API
will keep the client connected forever if it does not raise a 500 error soon
enough.

This patches make sure that the only part retrying for ever is metricd: the
rest (e.g. the API) will fail fast if anything bad happens.

Fixes #194