Create change
GitHub Pull Request: https://api.github.com/repos/gnocchixyz/gnocchi/pulls/145

metricd: use a better number of partitions for the hashring

The current hashring distributing the sacks across the metricd workers is 200.

That means the the smallest number of sacks that a worker can have is

(number_of_sacks/200), being 0.64 on a default installation.



That's weird ratio.



If the number of sacks created is 4096 and the number of workers is 4096, then

each worker manages at least 4096/200=20,48 sacks. There's no good reason for

that, each worker should be responsible for one sack (+ replicas) in such a

setup.



Let's use the number of sacks as the partition number so the minimum amount of

sack that a worker will always be able to manage is 1 (+ replicas).

Patch-set: 1
Change-id: I2b2068e3b64831733e3a3d31531f2388b37ed5d6
Subject: metricd: use a better number of partitions for the hashring
Branch: refs/heads/master
Commit: dc7e9839c674404d68d15545387f38225f04fcd8
Tag: autogenerated:gerrit:newPatchSet
Groups: dc7e9839c674404d68d15545387f38225f04fcd8
Work-in-progress: false