Merge gnocchi.storage._carbonara into gnocchi.storage
The plan was to keep the storage drivers Carbonara agnostic, but we cannot say
we succeeded. There are many places from metricd to gnocchi.storage where the
abstraction leaks and a lot of assumption about how driver works are based on
Carbonara.

This patch simplifies the code base by breaking the assumption that someone
will someday built a storage driver that is not based on Carbonara. I don't
think it will ever happen and that there is any point at this stage of the project.