ceph: order unprocessed measures before processing
This change adjusts the way in which Gnocchi with Ceph storage obtains
unprocessed measures for processing. Previously unprocessed measures can
enter a condition where specific metrics are favored which causes some
metrics with unprocessed measures to continuously grow. This continuous
growth of the backlog stored in a Ceph omap object in the form of keys
and a small 16-byte object will eventually cause an OSD to fail. By
sorting the order of unprocessed measures by date timestamp, we gain
three things:

1. Efficiency in that once "realtime" capacity of Gnocchi is exceeded, more
   measures are processed per processing of a metric
2. No more starved metrics, which grow unbounded in backlog until a failure
3. Scalability of Gnocchi before failures is extended above "realtime"
   processing capacity