storage: fix metric processing scheduling inefficiency
metricd scheduler learns about which metrics to process by calling
`list_metric_with_measures_to_process(size, part)`.

The problem currently is that list_metric_with_measures_to_process() uses
(size, part) to poke holes into a list of _measures_, which are very likely to
be ordered by the name of the measure file (true at least for file and ceph
drivers). That means that if the backlog is something like:

[ metric1_measure1, metric1_measure2, … metric2_measure1, metric2_measure2, …]

and if metric1 has e.g. 100k measures to be processed, then all schedulers will
poke holes starting from the beginning of the backlog:

 scheduler1 backlog:
  [ metric1_measure1, metric1_measure2, … metric1_measure256 ]

 scheduler2 backlog:
  [ metric1_measure257, metric1_measure258, … metric1_measure512 ]

which is then boiled down to:

  scheduler1 metric to process: metric1
  scheduler2 metric to process: metric1

this is completely inefficient as they might be tons of metrics to process but
all scheduler are focusing on the same first metric.

This patch tries to fix that by using the full measure listing, boiling down
the list of metric that needs to be processed, sorting it, and *then* poking
holes in that backlog (so the backlog is not about measures, but about metrics
this time). This is way more efficient as it would solve the case described
above.

The only downside is that the Swift driver, which was the only one not doing a
full measure listing on each call, has now to do a full listing too. But it
should be a trade-off worth it.