Add notification mechanism in incoming driver to be used by metricd worker
This allows for the incoming driver to provide a method that will iter on sacks
that have new measures to process. Metricd worker will wakeup earlier and start
processing sack.

This avoids polling randomly sacks to see if new measures are available, and
focus on sacks that have the most measures to process first.