Move cross aggregation computing to gnocchi.rest
This moves the cross aggregation of timeseries to a new dedicated
module gnocchi.rest.cross_metric. It turns out that this is one of the
latest module depending on Pandas (for probably good reasons).

However it is not directly related to Carbonara, so there's no need to
move it there. It could be put into gnocchi.storage, but it actually
have nothing to do with storage: it's only recomputing stuff once they
are retrieved.

Moving this to the REST area makes sure that it is only loaded by the
REST API server, and therefore prevent other processes such as metricd
to load Pandas (which is huge in memory).

Closes #61