Simplify error handling in IncomingDriver.get_storage_sacks
In some race conditions, the get_storage_sacks method gets called too
early. And exceptions were caught in get_storage_sacks and preventing
callers to generate meaningful logs.

get_storage_sacks is renamed to _get_storage_sacks to stress that it
is reserved to driver implementations. To retrieve the number of
storage sacks, the NUM_SACKS property is preferred.