metricd: fix coordinator not implemented error
If the coordinator does not implement partitioning, then a message is logged at
startup, but each subsequent called to run_watchers will raise a
tooz.NotImplemented. In that case an empty error message is logged, which is
scary. Just ignore that, since the "problem" is already being warned to the
user at startup anyway.