redis: do not emit notifications until sack is processed
The current behavior is to notify no matter what as soon as a new measure is
push into a metric, and therefore, into a sack.

This patches modifies this behaviour by having the measure storage putting a
flag into a special key named after the sack it is pushing into. Metricd
processes watches for this key to be set and get a notification _only the first
time it is set_. Once metricd finishes processing that sack, it delete the key
so it can get notified again if any new measures has arrived.

This patch also simplifies the notification processing in metricd by just
setting back on the notification system when processing is finished. Previous
version was trying to re-enable notification prior to processing but it just
complicates the workflow for little gain; there's no lock around notification,
and would just add overhead.