metricd: cache sacks to process
With the new notification system calling _get_sack_to_process as often as each
notification is received, the tooz `run_watchers` function is called too often,
hitting the backend a lot, often for no change.

This patch restores back the previous behavior of updating the list of sack to
process every `metric_processing_delay` maximum. In order to do that, it simply
leverages cachetools and cache the output for a defined TTL equals to
`metric_processing_delay`. That avoids trying to be too smart about when/how to
call run_watchers while giving the wanted result.

Closes #406