incoming: fix logging on SackDetectionError
Accessing NUM_SACKS and having SackDetectionError is the mechanism that
upgrade() uses to detect if it needs to do its job. But since logging is down
too low-level, the error message is print on the first call of gnocchi-upgrade,
which does not make any sense. Log only where it makes sense.