aggregates: make aggregated() extensible
This change moved out two things from aggreated():
* What aggregation we retrieve for each metrics from the backend
* What we do with the matrix that contains all points of all
  retrieved measures (instead of just reaggregating).

To do so, metrics become metrics_and_aggregation that is a list metric
with the aggregation to retrieve. And reaggregation become the more
generic operations. Operations is a list of transformation to do
on the measures matrix. This change only implements the reaggregation
as operations for now to replace "reaggregation". It looks like:
["aggregate", "mean", ["metric", ["id1", "mean"], ["id2", "mean"]]]

More operations will be added in future patches.

This doesn't add any tests since it's just refactoring. The new
operations mechanism is already excercise by the test_aggregates.py
file.

Related #419