api: convert exceptions JSON data struct rather than just a string
Currently all exceptions used in gnocchi.rest.api.abort() are serialized using
str(), which does not make them easy to use to consumer.

This cannot be changed as it'd break the API, but this adds a new mechanism so
new exceptions that are added can be serialized with a complete data structure
easy for computers to consume, rather than a string to parse.