simplify operations interface
it's very cluttered to pass around is_aggregated when it's only
ever used for end result formatting.

we'll only ever have two scenarios:
- everything is aggregated into a single result
- batch retrieval and return all series individually (maybe modified)

there doesn't seem to be path which has a partially aggregated
result. if there is, it doesn't seem like this code would even handle
it.

this patch:
- removes is_aggregated and just assumes, if there's one series
returned and there was more than one given, it's aggregated.
- just transpose once if multiple series returned.