aggregates: rework API output
This design have some issue when two timeseries have the same names,
only the last one is dumped. Or if the same metric is asked
twice for whatever reason, measures have all timestamps duplicated.

This change the result to a new format:

For resources:
{
  "measures": { resource_id: {metric_name: {aggregation: MEASURES}}},
  "references": [RESOURCES_DETAILS]
}

For metrics
{
  "measures": {metric_id: {aggregation: MEASURES}},
  "references": [METRICS_DETAILS]
}

When the result is aggregtated we got:

For resources:
{
  "measures": { "aggregated": MEASURES},
  "references": [RESOURCES_DETAILS]
}
For metrics
{
  "measures": { "aggregated": MEASURES},
  "references": [METRICS_DETAILS]
}

Closes-bug: #477