DNM: rest: port /v1/metric API to Falcon
This is a PoC that I made in a few hours to port the metric API to Falcon. I
wanted to have a glimpse on how easy it would be – and it was very easy.

For now, I like how easier the routing is and how cleaner and easier to grasp
it is compared to Pecan. However, I don't see any killer feature yet.

This patch passes all the Gabbi test, and some of the test_rest tests. I don't
see a point yet in making test_rest pass, as it is good enough for what I was
trying to do.