storage: stop checking for archive policy aggregation methods match
The storage API is currently responsible to check that requesting an
aggregation method for a metric is possible or not based on its archive policy.
This is out of its job actually, and it should just return an empty series if
it has nothing – or raise MetricDoesNotExist if the metric is unknown.

The API will take the job of doing this validation before hitting the storage
subsystem.