Retry and log on drivers initialization failure
This retries and logs all failure to connect to any of the driver (index,
storage, incoming and coordinator). The wait time is not changed (exponential
up to 1 minute), but now both the API and metricd uses it this mechanism.

The error is also logged, so that any failure that is not connection related
(e.g. Python module missing) is properly logged and the operator knows what he
going wrong:

2017-12-15 10:37:49,075 [9070] ERROR    gnocchi.utils: Unable to initialize coordination driver
Traceback (most recent call last):
  File "/Users/jd/Source/tenacity/tenacity/__init__.py", line 298, in call
    result = fn(*args, **kwargs)
  File "./gnocchi/cli/metricd.py", line 44, in get_coordinator_and_start
    coord.start(start_heart=True)
  File "/Users/jd/Source/tooz/tooz/coordination.py", line 423, in start
    self._start()
  File "/Users/jd/Source/tooz/tooz/drivers/mysql.py", line 139, in _start
    self._options)
  File "/Users/jd/Source/tooz/tooz/drivers/mysql.py", line 198, in get_connection
    cause=e)
  File "/Users/jd/Source/tooz/tooz/utils.py", line 225, in raise_with_cause
    excutils.raise_with_cause(exc_cls, message, *args, **kwargs)
  File "/usr/local/opt/python/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/oslo_utils/excutils.py", line 143, in raise_with_cause
    six.raise_from(exc_cls(message, *args, **kwargs), kwargs.get('cause'))
  File "/usr/local/opt/python/Frameworks/Python.framework/Versions/2.7/lib/python2.7/site-packages/six.py", line 737, in raise_from
    raise value
ToozConnectionError: (2003, "Can't connect to MySQL server on 'localhost' ([Errno 61] Connection refused)")

Fixes #585