storage: truncate aggregated time series before storing
The current truncation code in AggregatedTimeSerie truncate on object
initialize only. That means that split up to POINTS_PER_SPLIT points are
stored, and that data is read and _truncated_ at this point.

When new points are added (existing merge), it's possible that the number of
points grows more than the number of points defined by the archive policy,
storing too many points. That's hidden because when re-read, the
AggregatedTimeSerie is truncated again.

However, that makes the storage driver store too much data.

This patches remove the max_size attribute from carbonara.AggregatedTimeSerie
and simply expose a truncate() method that can be used to truncate each split
based on the oldest point that is defined by the archive policy.