fix median aggregation
i mucked up #602 and the tests didn't find it. basically, the ordered
array i thought i build was not ordered by group as required to
make the logic of finding max and finding diff to middle, it was just
ordered with no regard to group.

this fixes it and expands the test so it's a bit more coverage.

this unfortunately is slower than what we merged. this is ~1.3x faster
unlike what i claimed before as ~1.9x faster