Introduce Aggregation object
This introduces a new Aggregation objects (a named tuple) that stores the
information needed to retrieve any aggregation from the storage driver.

The goal is make most methods accept this Aggregation object as argument
instead of a list of kwargs and to replace the archive policy usage in the
storage driver.