cleanup only when new object
currently, we scan for objects to cleanup or to rewrite. this is
unnecessary i/o as the policies requires, at least, 1hrs worth of
data (at high policy), and at lower policies requires days worth of
data. it does not make sense for us to check for objects to clean
for every measure.

this leverages the fact we can compute when we need to rewrite
objects without scanning by computing new split vs old split. using
this knowledge, we scan only if we know a new object is created.
only in this case will we check to see what objects need to be
deleted or rewritten.

this will still end up scanning sometimes even though no work is
needed for non-append drivers (see: not ceph) but this should lower
the false scans significantly.

Fixes #608