metricd: restart scheduler on group membership issue
In certain (bad) situation, Gnocchi coordination backend can be e.g.
restarted/changed under its feet. While metricd thinks it has joined the group,
it actually is not part of it so the `members.index` call raises `ValueError`.

Here, we exit the scheduler instead so it is restarted by cotyledon has a
chance to reconfigure itself.